requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
astroid==3.3.8
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
dateparser==1.2.0
dill==0.3.9
discord==2.3.2
discord.py==2.4.0
emoji==2.14.1
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
parsedatetime==2.6
platformdirs==4.3.6
propcache==0.2.1
pycparser==2.22
pylint==3.3.3
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
yarl==1.18.3
yt-dlp==2025.1.15
